@@ -8,21 +8,21 @@
     "analyze": "set NODE_ENV=production&&webpack -p --json | webpack-bundle-size-analyzer"
   },
   "devDependencies": {
-    "array.prototype.find": "2.0.4",
-    "array.prototype.findindex": "2.0.2",
-    "babel-core": "6.26.3",
-    "babel-loader": "8.0.4",
+    "@babel/core": "^7.1.6",
+    "@babel/preset-env": "^7.1.6",
+    "@babel/preset-react": "^7.0.0",
+    "@dnnsoftware/dnn-react-common": "^2.0.5",
+    "array.prototype.find": "^2.0.0",
+    "array.prototype.findindex": "^2.0.0",
+    "babel-loader": "^8.0.4",
     "babel-plugin-transform-object-assign": "6.22.0",
     "babel-plugin-transform-object-rest-spread": "6.26.0",
     "babel-plugin-transform-react-remove-prop-types": "0.4.20",
     "babel-polyfill": "6.26.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
+    "create-react-class": "^15.6.3",
     "css-loader": "1.0.1",
-    "dnn-global-styles": "0.0.5",
-    "dnn-webpack-externals": "1.1.2",
-    "es6-object-assign": "1.1.0",
-    "eslint": "5.8.0",
+    "es6-object-assign": "^1.0.3",
+    "eslint": "5.9.0",
     "eslint-loader": "2.1.1",
     "eslint-plugin-react": "7.11.1",
     "eslint-plugin-spellcheck": "0.0.11",
@@ -32,15 +32,19 @@
     "less": "3.8.1",
     "less-loader": "4.1.0",
     "moment": "^2.15.0",
+    "prop-types": "^15.6.2",
     "raw-loader": "0.5.1",
+    "react": "^16.6.3",
     "react-click-outside": "^3.0.1",
+    "react-dom": "^16.6.3",
     "react-hot-loader": "4.3.12",
-    "style-loader": "^0.23.1",
-    "throttle-debounce": "^2.0.1",
+    "style-loader": "0.23.1",
+    "throttle-debounce": "2.0.1",
     "url-loader": "1.1.2",
-    "webpack": "4.25.1",
-    "webpack-bundle-size-analyzer": "2.7.0",
-    "webpack-dev-server": "^3.1.10"
+    "webpack": "4.26.0",
+    "webpack-bundle-size-analyzer": "3.0.0",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "3.1.10"
   },
   "dependencies": {
     "react-widgets": "^4.4.6"
