@@ -3,101 +3,64 @@
   "version": "1.0.2",
   "private": true,
   "scripts": {
-    "build": "set NODE_ENV=production&&webpack -p",
+    "build": "set NODE_ENV=production&&webpack -p --progress",
     "webpack": "webpack-dev-server -d --port 8070 --hot --inline --content-base dist/ --history-api-fallback",
     "analyze": "set NODE_ENV=production&&webpack --config webpack.config.js --profile --json > stats.json&&webpack-bundle-analyzer stats.json"
   },
   "devDependencies": {
+    "@babel/core": "^7.1.6",
+    "@babel/preset-env": "^7.1.6",
+    "@babel/preset-react": "^7.0.0",
+    "@dnnsoftware/dnn-react-common": "^2.0.5",
     "array.prototype.find": "2.0.4",
     "array.prototype.findindex": "2.0.2",
-    "babel-core": "6.26.3",
-    "babel-loader": "8.0.4",
+    "babel-loader": "^8.0.4",
     "babel-plugin-transform-object-assign": "6.22.0",
     "babel-plugin-transform-object-rest-spread": "6.26.0",
     "babel-plugin-transform-react-remove-prop-types": "0.4.20",
     "babel-polyfill": "6.26.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
+    "create-react-class": "^15.6.3",
     "css-loader": "1.0.1",
-    "dnn-back-to-link": "0.0.5",
-    "dnn-button": "0.0.11",
-    "dnn-checkbox": "0.1.5",
-    "dnn-collapsible": "0.0.13",
-    "dnn-collapsible-row": "0.0.8",
-    "dnn-content-load-wrapper": "0.0.5",
-    "dnn-date-picker": "0.8.9",
-    "dnn-dropdown": "1.0.21",
-    "dnn-dropdown-with-error": "0.0.11",
-    "dnn-editable-field": "0.0.17",
-    "dnn-file-upload": "^0.1.22",
-    "dnn-folder-picker": "0.0.7",
-    "dnn-global-styles": "0.0.5",
-    "dnn-grid-cell": "0.0.6",
-    "dnn-grid-system": "0.0.7",
-    "dnn-input-group": "0.0.10",
-    "dnn-label": "0.0.10",
-    "dnn-modal": "0.0.5",
-    "dnn-multi-line-input": "1.0.9",
-    "dnn-multi-line-input-with-error": "0.1.5",
-    "dnn-page-picker": "0.0.24",
-    "dnn-pager": "0.0.18",
-    "dnn-permission-grid": "0.0.22",
-    "dnn-persona-bar-page": "0.0.6",
-    "dnn-persona-bar-page-body": "0.0.6",
-    "dnn-persona-bar-page-header": "0.0.33",
-    "dnn-radio-buttons": "0.0.14",
-    "dnn-scrollbar": "1.0.1",
-    "dnn-search-box": "1.0.9",
-    "dnn-searchable-tags": "0.0.9",
-    "dnn-select": "0.0.8",
-    "dnn-single-line-input": "1.0.7",
-    "dnn-single-line-input-with-error": "0.1.5",
-    "dnn-slider": "0.0.4",
-    "dnn-sortable": "0.0.7",
-    "dnn-svg-icons": "0.1.22",
-    "dnn-switch": "0.1.6",
-    "dnn-tabs": "0.0.9",
-    "dnn-tags": "0.0.35",
-    "dnn-text-overflow-wrapper": "0.0.11",
-    "dnn-tooltip": "0.2.3",
-    "dnn-tree-control-interactor": "0.1.7",
     "es6-object-assign": "1.1.0",
     "es6-promise": "4.2.5",
-    "eslint": "5.8.0",
+    "eslint": "5.9.0",
     "eslint-loader": "2.1.1",
     "eslint-plugin-react": "7.11.1",
     "eslint-plugin-spellcheck": "0.0.11",
     "file-loader": "2.0.0",
+    "json-loader": "^0.5.7",
     "jwt-decode": "2.2.0",
     "less": "3.8.1",
     "less-loader": "4.1.0",
     "moment": "^2.22.1",
     "pikaday": "1.8.0",
+    "prop-types": "^15.6.2",
     "raw-loader": "0.5.1",
-    "react": "16.6.1",
+    "react": "^16.6.3",
     "react-collapse": "4.0.3",
     "react-custom-scrollbars": "4.2.1",
-    "react-dom": "16.6.1",
+    "react-dom": "^16.6.3",
     "react-height": "3.0.0",
     "react-hot-loader": "4.3.12",
     "react-modal": "3.6.1",
     "react-motion": "0.5.2",
-    "react-redux": "5.1.0",
+    "react-redux": "5.1.1",
     "react-tabs": "2.3.0",
     "react-tooltip": "3.9.0",
     "react-widgets": "4.4.6",
     "redux": "4.0.1",
-    "redux-devtools": "3.4.1",
+    "redux-devtools": "3.4.2",
     "redux-devtools-dock-monitor": "1.1.3",
     "redux-devtools-log-monitor": "1.4.0",
     "redux-immutable-state-invariant": "2.1.0",
     "redux-thunk": "2.3.0",
     "source-map-loader": "^0.2.3",
-    "style-loader": "^0.23.1",
-    "throttle-debounce": "^2.0.1",
+    "style-loader": "0.23.1",
+    "throttle-debounce": "2.0.1",
     "url-loader": "1.1.2",
-    "webpack": "4.25.1",
+    "webpack": "4.26.0",
     "webpack-bundle-analyzer": "^3.0.3",
-    "webpack-dev-server": "^3.1.10"
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "3.1.10"
   }
 }
