@@ -1,13 +1,12 @@
-import React, {Component, PropTypes } from "react";
+import React, {Component } from "react";
+import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import "./style.less";
 import "react-widgets/lib/less/react-widgets.less";
-import DropDown from "dnn-dropdown";
 import "./style.less";
 import resx from "../../../resources";
 import IconButton from "../../common/IconButton";
-import SearchBox from "dnn-search-box";
-import GridCell from "dnn-grid-cell";
+import { GridCell, SearchBox, Dropdown }  from "@dnnsoftware/dnn-react-common";
 import RoleGroupEditor from "../RoleEditor/RoleGroupEditor";
 import {
     roles as RolesActions
@@ -26,14 +25,6 @@ class FiltersBar extends Component {
             showPopup: false
         };
         canEdit = util.settings.isHost || util.settings.isAdmin || util.settings.permissions.EDIT;
-    }
-    componentWillMount() {
-    }
-    componentDidMount() {
-    }
-    componentWillUnmount() {
-    }
-    componentWillReceiveProps() {
     }
     onDeleteGroup() {
         const {props} = this;
@@ -63,9 +54,9 @@ class FiltersBar extends Component {
     }
     closeDropDown() {
         /*This is done in order to keep the dropdown closed on click on edit/delete*/
-        let {state} = this.refs["groupsDropdown"];
+        let {state} = this.groupsDropdownRef.current;
         state.dropDownOpen = false;
-        this.refs["groupsDropdown"].setState({
+        this.groupsDropdownRef.current.setState({
             state
         });
     }
@@ -131,13 +122,13 @@ class FiltersBar extends Component {
     getRoleGroupsDropDown() {
         let label = this.getCurrentLabel();
         let roleGroupsOptions = this.BuildRoleGroupsOptions();
-        let GroupsDropDown = <DropDown  style={{ width: "100%" }}
+        let GroupsDropDown = <Dropdown  style={{ width: "100%" }}
             withBorder={false}
             options={roleGroupsOptions}
             label={label }
             onSelect={this.onSelect.bind(this) }
-            ref="groupsDropdown"
-            />;
+            ref={node => this.groupsDropdownRef = node}
+        />;
         return <div className="groups-filter">{GroupsDropDown}<div className="clear"></div></div>;
     }
 
