@@ -5,41 +5,37 @@ const path = require("path");
 const isProduction = process.env.NODE_ENV === "production";
 module.exports = {
     entry: "./src/main.jsx",
+    optimization: {
+        minimize: isProduction
+    },
     output: {
-        path: "../admin/personaBar/scripts/bundles/",
+        path: path.resolve("../admin/personaBar/scripts/bundles/"),
         filename: "roles-bundle.js",
         publicPath: isProduction ? "" : "http://localhost:8080/dist/"
     },
 
     module: {
-        loaders: [
-            { test: /\.(js|jsx)$/, exclude: /node_modules/, loaders: ["react-hot-loader", "babel-loader"] },
-            { test: /\.less$/, loader: "style-loader!css-loader!less-loader" },
+        rules: [
+            { test: /\.(js|jsx)$/, enforce: "pre", exclude: /node_modules/, loader: "eslint-loader", options: { fix: true } },
+            { test: /\.(js|jsx)$/ , exclude: /node_modules/, loader: "babel-loader" },
+            { test: /\.(less|css)$/, loader: "style-loader!css-loader!less-loader" },
             { test: /\.(ttf|woff)$/, loader: "url-loader?limit=8192" },
-            { test: /\.css$/, loader: "style-loader!css-loader" },
             { test: /\.(gif|png)$/, loader: "url-loader?mimetype=image/png" },
             { test: /\.woff(2)?(\?v=[0-9].[0-9].[0-9])?$/, loader: "url-loader?mimetype=application/font-woff" },
             { test: /\.(ttf|eot|svg)(\?v=[0-9].[0-9].[0-9])?$/, loader: "file-loader?name=[name].[ext]" },
-        ],
-
-        preLoaders: [
-            { test: /\.(js|jsx)$/, exclude: /node_modules/, loader: "eslint-loader" }
         ]
     },
-
     resolve: {
-        extensions: ["", ".js", ".json", ".jsx"],
-        root: [
-            path.resolve('./src'),          // Look in src first
-            path.resolve('./node_modules')  // Last fallback to node_modules
+        extensions: [".js", ".json", ".jsx"],
+        modules: [
+            path.resolve("./src"),          // Look in src first
+            path.resolve("./node_modules")  // Last fallback to node_modules
         ]        
     },
 
-    externals: require("dnn-webpack-externals"),
+    externals: require("@dnnsoftware/dnn-react-common/WebpackExternals"),
 
     plugins: isProduction ? [
-        new webpack.optimize.UglifyJsPlugin(),
-        new webpack.optimize.DedupePlugin(),
         new webpack.DefinePlugin({
             VERSION: JSON.stringify(packageJson.version),
             "process.env": {
@@ -47,8 +43,8 @@ module.exports = {
             }
         })
     ] : [
-            new webpack.DefinePlugin({
-                VERSION: JSON.stringify(packageJson.version)
-            })
-        ]
+        new webpack.DefinePlugin({
+            VERSION: JSON.stringify(packageJson.version)
+        })
+    ]
 };
