@@ -128,5 +128,23 @@ module.exports = [
     "devtools",
     "ctrl",
     "mixin",
-    "collapsable"
+    "collapsable",
+    "Polyfills",
+    "findindex",
+    "adler32",
+    "Namespaces",
+    "Fn",
+    "Accessor",
+    "Reconciler",
+    "Devtool",
+    "Textarea",
+    "Validator",
+    "Checksum",
+    "Perf",
+    "Renderer",
+    "Subtree",
+    "Responder",
+    "Viewport",
+    "jsx",
+    "interactor"
 ]
